# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CertificateMetadataResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kms_provider": {
            "title": "Kms Provider",
            "type": "string"
          },
          "nombre_titular": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Titular"
          },
          "revocado_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revocado At"
          },
          "rut_titular": {
            "title": "Rut Titular",
            "type": "string"
          },
          "storage_provider": {
            "title": "Storage Provider",
            "type": "string"
          },
          "valido_desde": {
            "title": "Valido Desde",
            "type": "string"
          },
          "valido_hasta": {
            "title": "Valido Hasta",
            "type": "string"
          }
        },
        "required": [
          "id",
          "rut_titular",
          "nombre_titular",
          "valido_desde",
          "valido_hasta",
          "revocado_at",
          "kms_provider",
          "storage_provider"
        ],
        "title": "CertificateMetadataResponse",
        "type": "object"
      },
      "CertificateUploadRequest": {
        "additionalProperties": false,
        "properties": {
          "nombre_titular": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Titular"
          },
          "passphrase": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passphrase"
          },
          "pfx_base64": {
            "minLength": 4,
            "title": "Pfx Base64",
            "type": "string"
          },
          "rut_titular": {
            "maxLength": 12,
            "minLength": 3,
            "title": "Rut Titular",
            "type": "string"
          },
          "valido_desde": {
            "format": "date",
            "title": "Valido Desde",
            "type": "string"
          },
          "valido_hasta": {
            "format": "date",
            "title": "Valido Hasta",
            "type": "string"
          }
        },
        "required": [
          "pfx_base64",
          "rut_titular",
          "valido_desde",
          "valido_hasta"
        ],
        "title": "CertificateUploadRequest",
        "type": "object"
      },
...
      "CreateEmpresaRequest": {
        "additionalProperties": false,
        "properties": {
          "capital_inicial_uf": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capital Inicial Uf"
          },
          "fecha_inicio_actividades": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Inicio Actividades"
          },
          "giro": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Giro"
          },
          "razon_social": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Razon Social",
            "type": "string"
          },
          "regimen_actual": {
            "default": "desconocido",
            "enum": [
              "14_a",
              "14_d_3",
              "14_d_8",
              "presunta",
              "desconocido"
            ],
            "title": "Regimen Actual",
            "type": "string"
          },
          "rut": {
            "maxLength": 12,
            "minLength": 3,
            "title": "Rut",
            "type": "string"
          }
        },
        "required": [
          "rut",
          "razon_social"
        ],
        "title": "CreateEmpresaRequest",
        "type": "object"
      },
...
      "EmpresaResponse": {
        "properties": {
          "capital_inicial_uf": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capital Inicial Uf"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "fecha_inicio_actividades": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Inicio Actividades"
          },
          "giro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Giro"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "razon_social": {
            "title": "Razon Social",
            "type": "string"
          },
          "regimen_actual": {
            "enum": [
              "14_a",
              "14_d_3",
              "14_d_8",
              "presunta",
              "desconocido"
            ],
            "title": "Regimen Actual",
            "type": "string"
          },
          "rut": {
            "title": "Rut",
            "type": "string"
          }
        },
        "required": [
          "id",
          "rut",
          "razon_social",
          "giro",
          "regimen_actual",
          "fecha_inicio_actividades",
          "capital_inicial_uf",
          "created_at"
        ],
        "title": "EmpresaResponse",
        "type": "object"
      },
      "EmpresasListResponse": {
        "properties": {
          "empresas": {
            "items": {
              "$ref": "#/components/schemas/EmpresaResponse"
            },
            "title": "Empresas",
            "type": "array"
          }
        },
        "required": [
          "empresas"
        ],
        "title": "EmpresasListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MandatoCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "alcance": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Alcance",
            "type": "array"
          },
          "consentimiento_version": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Consentimiento Version",
            "type": "string"
          },
          "inicio": {
            "format": "date",
            "title": "Inicio",
            "type": "string"
          },
          "ip_otorgamiento": {
            "anyOf": [
              {
                "maxLength": 45,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Otorgamiento"
          },
          "sii_referencia": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sii Referencia"
          },
          "termino": {
            "format": "date",
            "title": "Termino",
            "type": "string"
          }
        },
        "required": [
          "alcance",
          "inicio",
          "termino",
          "consentimiento_version"
        ],
        "title": "MandatoCreateRequest",
        "type": "object"
      },
      "MandatoResponse": {
        "properties": {
          "alcance": {
            "items": {
              "type": "string"
            },
            "title": "Alcance",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "inicio": {
            "title": "Inicio",
            "type": "string"
          },
          "revocado_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revocado At"
          },
          "sii_referencia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sii Referencia"
          },
          "termino": {
            "title": "Termino",
            "type": "string"
          }
        },
        "required": [
          "id",
          "alcance",
          "inicio",
          "termino",
          "revocado_at",
          "sii_referencia"
        ],
        "title": "MandatoResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Renteo API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/empresas": {
      "get": {
        "description": "Lista empresas visibles bajo RLS, ordenadas por created_at desc.",
        "operationId": "list_empresas_api_empresas_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresasListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Empresas",
        "tags": [
          "empresas"
        ]
      },
      "post": {
        "operationId": "create_empresa_api_empresas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmpresaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Empresa",
        "tags": [
          "empresas"
        ]
      }
    },
...
    "/api/empresas/{empresa_id}/certificado": {
      "delete": {
        "operationId": "revoke_certificate_endpoint_api_empresas__empresa_id__certificado_delete",
        "parameters": [
          {
            "in": "path",
            "name": "empresa_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Empresa Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Certificate Endpoint",
        "tags": [
          "custodia"
        ]
      },
      "get": {
        "operationId": "get_certificate_api_empresas__empresa_id__certificado_get",
        "parameters": [
          {
            "in": "path",
            "name": "empresa_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Empresa Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateMetadataResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Certificate",
        "tags": [
          "custodia"
        ]
      },
      "post": {
        "operationId": "upload_certificate_api_empresas__empresa_id__certificado_post",
        "parameters": [
          {
            "in": "path",
            "name": "empresa_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Empresa Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CertificateUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateMetadataResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Certificate",
        "tags": [
          "custodia"
        ]
      }
    },
    "/api/empresas/{empresa_id}/mandato": {
      "get": {
        "operationId": "get_mandato_api_empresas__empresa_id__mandato_get",
        "parameters": [
          {
            "in": "path",
            "name": "empresa_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Empresa Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MandatoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Mandato",
        "tags": [
          "custodia"
        ]
      },
      "post": {
        "operationId": "register_mandato_api_empresas__empresa_id__mandato_post",
        "parameters": [
          {
            "in": "path",
            "name": "empresa_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Empresa Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MandatoCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MandatoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Mandato",
        "tags": [
          "custodia"
        ]
      }
    },
...
  }
}